gather-player-data/pyproject.toml
[project]
name = "gather-player-names"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.45",
    "names-dataset>=3.3.1",
 "pandas>=2.3.1",
]
